Improve error handling
Instead of returning strings, now return a full ApiError that allows to
peak inside and even has the `Error` type present. The returned error
contains the statusCode and `Errors` (when possible) and a description
that comes from the schema in case the statusCode comes from the server,
the error message in case connection was made, or simply a `contact
support if it happens again` in case the problem seems to be at setting
up the connection (possible programming error).

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>