Fix ESM support and publish mechanism.
There were still issues with the ESM support and the publish step.
Mostly, the publish wasn't actually only publishing the dist folder, and
the ESM support wasn't working due to missing `.js` terminations on the
imports and exports. This should now correctly make the library working
and fully supporting ESM and CJS.
Additionally, the package.json file was wrong, since we want to use two
different ones, one for development and another for actually
distributing. Added the distributing one and fixed copy mechanism

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>