Add dist README
We do not want to present the same README in the github and npm. For the
npm, we're going with more focus on what it does and how to use, while
the github readme is much more about how the code is generated and
what's left to do on the project.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>