Allow client to provide interceptors to requests
To make it easier to handle the requests on the client side, allow for
interceptors to be added that can either change the request or the
response in a predetermined way. These are given at client creation time
and are optional.

This is a basic use case, we have in our current dashboard to do
camelCase -> snake_case conversion and viceversa. I'm not sure this is
needed on the dashboard in particular, but the functionality itself is
still valuable.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>