Conver models/schema parameters to camelCase
The API has parameters following the snake_case convention. This is the
usual scenario on API's so we want to keep. However, the TS/JS ecosystem
usually goes for camelCase. Hence, we change all parameters in models
and schemas to become camelCase.

This is done in a similar manner as fixing the services is, via seds on
those files.

At this point, I do agree that I might have gone a bit overboard with the
complexity introduced on the Makefile to produce the actual desired
format. It might have been more benefitial to modify the generator tool
to better suit our needs, but I wasn't expecting so many little changes
in my initial setup. For now, I think it's better to stick to the current
way and in the future we might move these functionalities to actually be
in the generation of the base models themselves.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>