Use lune fork for base generator
Instead of using the original openapi-typescript-generator package, we
want to use the custom fork we've created to be able to provide with
base models and services that better fit our needs. For now, the initial
version is an exact replicate of the original behaviour so no changes in
the code itself are made.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>