Upgrade to new base_client version (#16)
Use the newest version of the base client generator. This
version exports a full client we can use, fully customised to Lune
needs. Because of that, all the `sed | grep` operations we did to
build the final client are now gone. In a way, the whole process
of building the final client from a base one is gone, hence some
changes on the `docker-compose-ci` and `README`.

Additionally, this new base client version already contains the
models in a camelCase version, with an interceptor already
added to the client to convert the responses on the endpoints.
To convert when sending the request, to prevent having to
decode what was already encoded, this process has been
added to the core bits of the request creation. Due to this,
camelCase and snakeCase libraries are now required on the
library and `esModuleInterop` and
`allowSyntheticDefaultImports` had to be activated. Maybe
in the future we can improve on this, feels unnecessary to use
3rd party libraries for our use case but for now avoiding bigger
code introductions.

Finally, The previous service generation had a
`throws APIError` being documented on the services, which
was made untrue when we moved to `Result` but wasn't
removed from the docs. The new client generation has it fixed
so those changes appear here.

I've tested this locally, and it's fully working on response and
error 😄 

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>