Add helper methods for local development
Add a new command that builds from remote schema without installing
dependencies nor fixing linting. This is to be used to regenerate the
models from the remote schema in the dashboard in a development
environment.
Since this command is only to be used on development, it was added to
the dev docker-compose.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>