Refactor publish mechanism
In an attempt to streamline the deployment process, the cleanest way
seems to be always try and publish to npm on a push to master while
ignoring the publishing same version error. With the addition of a
dedicated workflow to up the patch version, now deploying the changes in
master is simply triggering a workflow and approving the generated PR.

The changes introduced are:
- Add a docker command to bump the version (patch).
- Add a new workflow that runs in an environment that creates a PR with
  a version bump (patch).
- Publish mechanism is now triggered by any changes to `package.json`
  instead of a new release. This is triggered by any version change but
also by any change in dependencies.
- Publish mechanism is now ignoring the error of trying to publish over
  the same version. This means that if the version hasn't changed,
publish won't occur but it won't fail.
- Added a new workflow (to be ran in an environment) that allows to
  create an automatic PR with a patch version bump. This makes it easier
for anyone to create a publish workflow without having to manually
create PR's.
- Updated README

In the end, the workflow to publish is now:
- Create PR with version increase. If patch, one can use the newly added
workflow via Github by just clicking a button.
- Once the PR is approved and merged, deploy will happen.

This effectively removes the releases from how we publish things to npm.
Since the releases were only used for triggering the act of deployment,
I think dropping them is better to ease the publish process.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>