Create workflow for automatic patch deployment
Create a workflow that increases the package version (patch), commits
this change to master and then creates a new tagged release with
automatic release notes based on it. This should then trigger the other
github workflow to actually publish the package.

The workflow is on an environment that requires manual approval, meaning
the action is behind an approval in github.

The purpose is to make it easier for anyone to publish a new patch
version of the library with a single click, instead of the current
process that requires checking the code, upping the version, creating
PR, merging and then creating the release.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>