Remove bump version (patch) workflow (#60)
The workflow is currently not working (due to permissions pushing to
master repo) and required changes in the repo permissions that are not
very good. The workflow also has an issue where it can trigger itself
(albeit still requiring manual intervention) due to the creation of a
new commit on master.

This will be replaced by an alternative way to publish to npm, instead
of having this job. More in the future.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>