Update client with schema and generator reworks
There has been a large rework on the api schema definition and on the
base client generator itself.

The schema definition changes are:
- `operationId` changes which makes this version of the client require
  changes in our clients since service method names might have changed.
- `error` definitions (error_code and message) make on the wire
  difference in the API, which means when that version is deployed to
the BE, the old client will have issues.
- `summary` changes only really change comments so this is not a
  breaking change.

The base client generator changes are:
- Use the new generator that differs in how it calls the `request`
  method. Instead of presenting the `requestBody` as an object and all
other parameters one by one, we now have `data` that unites
`requestBody` and `queryParams`. There is also the addition of an
`options` object that currently contains an accountId that can be used
to override `Lune-Account` header. Some changes in `core/request.ts`
have been made to accomodate the new usage.

These changes have not yet been merged in the BE schema, I've
generated the changes locally and have set it as the target schema for
generation. This makes it possible to review these changes while
things aren't merged and to allow `lune-dashboard` work to start as
well based on this branch (https://github.com/lune-climate/lune-dashboard/pull/521)
Whenever things are merged in the BE side, we can build as normally.

Also, this is a BIG change. Accompanied by the breaking BE changes
I'm pretty sure this should be the first case of a major version bump 😄

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>