Update client with latest base generator (#93)
The base client generator changes are:
- `request` method has a different interface. Instead of presenting the
`requestBody` as an object and all other parameters one by one, we
now have `data` that unites `requestBody` and `queryParams`. There
is also the addition of an `options` object that currently contains an
accountId that can be used to override `Lune-Account` header.
Changes in `core/request.ts` have been made to accommodate the
new usage.
- Removed `withAccount` in favour of `options` per endpoint. This
makes the operations thread safe.

Also, this is a BIG change that requires changes in our users. I think
this might be a case where we want a major version bump 😄

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>