Update to latest base generator
This has fixes for the allOf at the root level. Models were regenerated.

PS: I don't particularly like that we've broken compatibility with
previous versions. Even though the on wire representation is backwards
compatible, the way to use the library is now different so it will
require some user changes to accomodate. I'm wondering if we should
solve the issue differently on the base generator to keep backwards
compatibility, I'll attempt a few things.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>