Update to latest base generator (#244)
This has fixes for the allOf at the root level. Models were regenerated.

PS: I don't particularly like that we've broken compatibility with
previous versions. Even though the on wire representation is backwards
compatible, the way to use the library is now different so it will
require some user changes to accommodate. I'm wondering if we should
solve the issue differently on the base generator to keep backwards
compatibility, I'll attempt a few things.

EDIT: After some trials, keeping backwards compatibility for this case is
showing up to be quite a hard task to do, will probably require a big
change on the base generator and how it generates the services. There's
lots of edge cases that make things complicated:
- oneOf inside allOf: the properties need to go outside the data field since
it's a oneOf type, but the way service templates are this is quite a big
change
- allOf with queryParams and how it is presented to the user etc. It would
break our current agreement of presenting formData followed by
queryParams for example. Can be solved but once again a big change
considering the current status quo
- I'm sure I'm forgetting some other edge cases here (like allOf inside allOf).

Since the current solution correctly handles everything (I believe) I'm leaving it as is.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>