Reject this kind of error
I had a brief look at how Axios constructs AxiosError instances and from
what I saw the request may be missing only if the error is of a kind
related to us using Axios incorrectly.

Not only it doesn't make sense to convert something like that to
ApiError (as it has nothing to do with the actual API, we never managed
to even make a request via network) if we have an error of this sort
it'll be happening all the time and prevent the library from working at
all.

This change will allow me to refactor this module to make error handling
easier to the users of this library.