Model the API errors better
This way we know that we either have just a description (connection
error kind of error) or we have gotten a response, in which case we have
status code and likely request id and errors too.

The request id and errors are optional (for now) – our API or reverse
proxies between us and the customers don't necessarily return these in
all cases, better to be defensive here.