Adapt to recent error-related API changes
We've had an API specification change recently[1] that changed the
error-related types and, as such, we need to make some manual changes to
handle that.

The build (generating new lune-ts version based on the API changes) has
been broken since [1] and this change unbreaks it.

The change consists of regenerating the types from the OpenAPI
specification from [1] (excluding any subsequent changes, we'll merge
them separately once this change is in place) and adapting ApiError.ts.

We can depend on there being a single error now (the error property in
the response, not the errors array) so the client code error reporting
can make use of it.

[1] https://github.com/lune-climate/lune-backend/commit/5fb0c0b138623b6729bdf73bdd27b431a2a71bd5

Resolves: https://linear.app/lune/issue/LUN-2964/the-ts-client-fails-to-ingest-new-openapi-specification-changes