Client auto update (#566)
Due to the latest changes in the base generator we want to rebuild our
code. Keep in mind this wasn't done automatically on the latest auto
generated PR since the bumping of the version was not yet in place
before that one was merged. This is just making things in sync.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>