[DRAFT] Client with resolved legs API schema
This is only meant to be used internally and to unblock Charley on her
FE work. It's useful to have a client that uses the expected API changes
that will be done on the emission estimate response.

The main changes are:
- introduce `methodology_details` which provides all the information for
  the new flight methodology.
- add `resolved_legs` to each of the leg responses. The resolved leg is
  pretty much the same as the input leg.

The client was generated via a local schema, the Makefile was slightly
adapted to use it instead of the remote one. This exact schema is fully
backwards compatible, the BE builds using the current schema :)

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>