Add multiple API versions capability
Our new base generator has introduced [^1] capabilities to allow using a
specific API version via the `options` field. This adapts the library to
correctly use this new option.

Keep in mind that there will be more changes introduced when the base
generator is shipped, the dependency here upgraded and everything
regenerated. For now the PR contains the required changes on the actual
library to allow the new field to be used.

[^1]: https://github.com/lune-climate/openapi-typescript-codegen/pull/73

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>