Upgrade base generator
Rework the logic and integrate the latest changes on the base generator.
As discussed in the PR, we moved to instead having `lune-ts` tied to a
specific API version, that is determined when we choose to update from
the remote schema. We now visit the changelog to understand the schema
version.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>