Adapt build process to new docs
We extract the current version information from our public changelog.
Our docs are changing in a bit and our changelog page will not be
presented for now. Since we do not use API versions for now, let us
hardcode the version. Whenever the public changelog page is back we can
bring the previous code or alternatively have a better way of knowing
the latest calendar version.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>