Halt build of API version fetching fails (#818)
We fetch the current API version from the changelog that is publicly
available in our docs. This step however can fail. Since the API version
is a requirement to building the application, we should fail if the
fetch has failed.

This was triggered due to an auto update PR which showcased an API
version of `<html>` in a build due to a failure in fetching the
changelog.

[^1]: https://github.com/lune-climate/lune-ts/pull/817

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>