Modernize the dockerized Node/npm setup (#977)
Just keeping things up to date and making sure we don't
fall too much behind the latest developments.

It all starts with the latest Node version which, combined
with no-longer-downgraded npm version allow us to get rid of
a series of hacks in the Docker image build process *I hope*
(why I hope: the build succeeds in this branch but it's
possible some other job fails after this is merged as some
code only runs in the master branch).

We have to add --legacy-peer-deps because newer npm refuses
to install invalid peer dependencies (good) so we have to
silence it until we resolve all of that.